requirements.txt
flask
textstat
pdfminer.six
